Fix duration calculation and timer precision

+++ MicroCode/src/timer.c
@@ -4,7 +4,7 @@ static void (*intrpt_ptr2)(); //only accessible from file
 static void (*intrpt_ptr3)();
 static void (*intrpt_ptr4)();
 
-void initTimer(TIM_TypeDef* tim, uint32_t rcc, uint32_t irq, uint16_t prescaler, uint16_t reloadValue) {
+void initTimer(TIM_TypeDef* tim, uint32_t rcc, uint32_t irq, uint16_t prescaler, uint32_t reloadValue) {
 	// Send power to the timer
     RCC->APB1ENR |= rcc;
 
@@ -32,17 +32,17 @@ void initTimer(TIM_TypeDef* tim, uint32_t rcc, uint32_t irq, uint16_t prescaler,
     NVIC_SetPriority(irq, 1);
 }
 
-void initTimer2(uint16_t prescaler, uint16_t reloadValue, void (*fun_ptr)()) {
+void initTimer2(uint16_t prescaler, uint32_t reloadValue, void (*fun_ptr)()) {
     intrpt_ptr2 = fun_ptr;
     initTimer(TIM2, RCC_APB1Periph_TIM2, TIM2_IRQn, prescaler, reloadValue);
 }
 
-void initTimer3(uint16_t prescaler, uint16_t reloadValue, void (*fun_ptr)()) {
+void initTimer3(uint16_t prescaler, uint32_t reloadValue, void (*fun_ptr)()) {
     intrpt_ptr3 = fun_ptr;
     initTimer(TIM3, RCC_APB1Periph_TIM3, TIM3_IRQn, prescaler, reloadValue);
 }
 
-void initTimer4(uint16_t prescaler, uint16_t reloadValue, void (*fun_ptr)()) {
+void initTimer4(uint16_t prescaler, uint32_t reloadValue, void (*fun_ptr)()) {
     intrpt_ptr4 = fun_ptr;
     initTimer(TIM4, RCC_APB1Periph_TIM4, TIM4_IRQn, prescaler, reloadValue);
 }
